// src/models/address_book.model.ts
export interface ADDRESS_BOOK  {
    address_book_id?:  number;
    customers_id:  number;
    entry_gender:  string;
    entry_company:  string;
    entry_firstname:  string;
    entry_lastname:  string;
    entry_street_address:  string;
    entry_suburb:  string;
    entry_postcode:  string;
    entry_city:  string;
    entry_state:  string;
    entry_country_id:  number;
    entry_zone_id:  number
}